+++ src/app/components/admin/department/department.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { DepartmentModel } from 'src/app/models/department.model';
 import { DepartmentRestService } from 'src/app/services/departmentRest/department-rest.service'; 
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-department',
@@ -37,12 +38,22 @@ export class DepartmentComponent implements OnInit {
   saveDepartment(saveDepartmentForm: any){
     this.departmentRest.saveDepartment(this.department).subscribe({
       next: (res:any)=>{
-        alert(res.message)
+        Swal.fire({
+          icon: 'success',
+          title: res.message,
+          showConfirmButton:false,
+          timer: 1300,
+        });
         this.getDepartments()
         saveDepartmentForm.reset();
       },
       error: (err)=>{
-        alert(err.error.message || err.error)
+        Swal.fire({
+          icon: 'warning',
+          title: err.error.message || err.error,
+          showConfirmButton:false,
+          timer: 1300,
+        });
       }
     })
   }
@@ -50,12 +61,22 @@ export class DepartmentComponent implements OnInit {
   updateDepartment(){
     this.departmentRest.updateDepartment(this.departmentGetId,this.departmentGetId._id).subscribe({
       next: (res:any)=>{
-        console.log(this.departmentGetId);
+        Swal.fire({
+          icon: 'success',
+          title: res.message,
+          showConfirmButton:false,
+          timer: 1300,
+        });
         
         this.getDepartments()
       },
       error: (err)=>{
-        alert(err.error.message || err.error)
+        Swal.fire({
+          icon: 'warning',
+          title: err.error.message || err.error,
+          showConfirmButton:false,
+          timer: 1300,
+        });
   
       }
     })
@@ -76,16 +97,40 @@ export class DepartmentComponent implements OnInit {
 }
 
 deleteDepartment(id:string){
-  this.departmentRest.deleteDepartment(id).subscribe({
-    next: (res:any)=>{
-      alert(res.message)
-      this.getDepartments()
-    },
-    error: (err)=>{
-      alert(err.error.message || err.error)
-
+  Swal.fire({
+    title: '¿Quieres eliminar el departamento',
+    showDenyButton: true,
+    showCancelButton: false,
+    confirmButtonText: 'Eliminar',
+    denyButtonText: `No eliminar`,
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+      this.departmentRest.deleteDepartment(id).subscribe({
+        next: (res:any)=>{
+          Swal.fire({
+            icon: 'success',
+            title: res.message,
+            showConfirmButton:false,
+            timer: 1300,
+          });
+          this.getDepartments()
+        },
+        error: (err)=>{
+          Swal.fire({
+            icon: 'warning',
+            title: err.error.message || err.error,
+            showConfirmButton:false,
+            timer: 1300,
+          });
+    
+        }
+      })
+    } else if (result.isDenied) {
+      Swal.fire('No se ha eliminado el departamento')
     }
   })
+
 }
 }
