+++ src/app/components/admin/destiny/destiny.component.ts
@@ -3,6 +3,7 @@ import { LodgeRestService } from 'src/app/services/lodgeRest/lodge-rest.service'
 import { DestinyRestService } from 'src/app/services/destinyRest/destiny-rest.service';
 import { TripRestService } from 'src/app/services/tripRest/trip-rest.service';
 import { DestinyModel } from 'src/app/models/destiny.model';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-destiny',
@@ -72,13 +73,23 @@ export class DestinyComponent implements OnInit {
   updateDestiny(){
     this.destinyRest.updateDestinyAdmin(this.destinyGetId,this.destinyGetId._id).subscribe({
       next: (res:any)=>{
-        alert(res.message)
+        Swal.fire({
+          icon: 'success',
+          title: res.message,
+          showConfirmButton:false,
+          timer: 1300,
+        });
         this.getLodges();
         this.getTrips();
         this.getDestinys();
       },
       error: (err)=>{
-        alert(err.error.message || err.error)
+        Swal.fire({
+          icon: 'warning',
+          title: err.error.message || err.error,
+          showConfirmButton:false,
+          timer: 1300,
+        }); 
   
       }
     })
@@ -98,17 +109,47 @@ export class DestinyComponent implements OnInit {
 }
 
 deleteDestiny(id:string){
-  this.destinyRest.deleteDestinyAdmin(id).subscribe({
-    next: (res:any)=>{
-      alert(res.message)
 
-      this.getDestinys();
-    },
-    error: (err)=>{
-      alert(err.error.message || err.error)
-
+  Swal.fire({
+    title: '¿Quieres eliminar el destino?',
+    showDenyButton: true,
+    showCancelButton: false,
+    confirmButtonText: 'Eliminar',
+    denyButtonText: `No eliminar`,
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+     
+      this.destinyRest.deleteDestinyAdmin(id).subscribe({
+        next: (res:any)=>{
+          Swal.fire({
+            icon: 'success',
+            title: res.message,
+            showConfirmButton:false,
+            timer: 1300,
+          });
+    
+          this.getDestinys();
+        },
+        error: (err)=>{
+          Swal.fire({
+            icon: 'warning',
+            title: err.error.message || err.error,
+            showConfirmButton:false,
+            timer: 1300,
+          }); 
+    
+        }
+      })
+    
+    } else if (result.isDenied) {
+      Swal.fire('No se ha eliminado el destino')
     }
   })
+
+
+
+
 }
 
 }
